camera_feed_api/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
yt-dlp==2025.07.01
python-multipart==0.0.6
google-genai==0.16.0